Add graphql trigger subscription
