+++ Yousei/Api/Types/FlowConfigInput.cs
@@ -7,17 +7,6 @@ using Yousei.Shared;
 
 namespace Yousei.Api.Types
 {
-    public record BlockConfigInput(string Type, string Configuration = "default", JToken arguments = default)
-    {
-        public BlockConfig Map()
-            => new()
-            {
-                Type = Type,
-                Configuration = Configuration,
-                Arguments = arguments,
-            };
-    }
-
     public record FlowConfigInput(BlockConfigInput Trigger, BlockConfigInput[] Actions = default)
     {
         public FlowConfig Map()
