+++ Yousei/Properties/launchSettings.json
@@ -2,11 +2,12 @@
   "profiles": {
     "Yousei": {
       "commandName": "Project",
-      "launchBrowser": false,
+      "launchBrowser": true,
       "environmentVariables": {
         "ASPNETCORE_ENVIRONMENT": "Development"
       },
-      "applicationUrl": "http://localhost:5000"
+      "applicationUrl": "http://localhost:5000/",
+      "launchUrl": "graphql"
     },
     "Docker": {
       "commandName": "Docker",
